training/requirements.txt
# Training dependencies
torch>=2
torchvision>=0.17
onnx>=1.16
onnxruntime>=1.18
onnxscript>=0.1
numpy>=1
Pillow>=10
